[APPDEV-2113] Real Time Dashboard
git-svn-id: https://appdevsvn.nj01/appdev/FreshDirect/trunk@16896 842219ff-9341-0410-a943-573460bdd4de

+++ projects/FDServices/src/com/freshdirect/analytics/model/RollData.java
@@ -4,7 +4,7 @@ import java.util.Date;
 
 public class RollData {
 
-	private int cnt;
+	private float cnt;
 	private String zone;
 	private String sector;
 	private Date cutOff;
@@ -31,10 +31,10 @@ public class RollData {
 	public void setSnapshotTimeFormatted(String snapshotTimeFormatted) {
 		this.snapshotTimeFormatted = snapshotTimeFormatted;
 	}
-	public int getCnt() {
+	public float getCnt() {
 		return cnt;
 	}
-	public void setCnt(int cnt) {
+	public void setCnt(float cnt) {
 		this.cnt = cnt;
 	}
 	public String getZone() {
